requirements.txt
mysql-connector-python==8.0.23
sendypy==5.0
